refactor(css): move component styles into src/css and update imports
