--- src/components/dashboard/cards/CardShared.css
@@ -1,38 +1,0 @@
-/* ============================================================
-   UNIVERSAL CARD HEADER + HELP ICON ALIGNMENT
-   Applies consistently to all dashboard cards:
-   - Price chart     (.card-title-row)
-   - Top signals     (.filters-card-header)
-   - Sentiment card  (.sentiment-card__header)
-   ============================================================ */
-
-.card-title-row,
-.filters-card-header,
-.sentiment-card__header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.75rem; /* uniform spacing across all card headers */
-  margin: 0;
-  padding: 0;
-}
-
-/* Headings inside headers align consistently */
-.card-title-row h2,
-.card-title-row h3,
-.filters-card-header h3,
-.sentiment-card__header h3 {
-  display: flex;
-  align-items: center;
-  gap: 0.45rem;  /* unified space between text + icon */
-  margin: 0;     /* remove default h2/h3 margin */
-  padding: 0;
-}
-
-/* HelpTooltip's trigger button */
-.card-title-row .help-icon-button,
-.filters-card-header .help-icon-button,
-.sentiment-card__header .help-icon-button {
-  margin: 0;
-  flex-shrink: 0; /* prevents icon from squishing on smaller screens */
-}
