--- src/components/dashboard/cards/StatSummary.css
@@ -1,63 +1,0 @@
-/* ========= STAT SUMMARY (top 3 cards) =============================== */
-
-.stat-row {
-  display: grid;
-  grid-template-columns: repeat(3, minmax(0, 1fr));
-  gap: 16px;
-  margin-bottom: 24px;
-}
-
-/* stat-card extends .panel from AppShell via JSX: className="panel stat-card" */
-.stat-card {
-  padding: 16px 18px; /* small tweak on top of .panel padding */
-}
-
-/* All stat cards share the same green accent at the top */
-.stat-card--accent-blue,
-.stat-card--accent-orange,
-.stat-card--accent-teal {
-  border-top: 4px solid var(--ustock-green-main);
-}
-
-.stat-label {
-  font-size: 11px;
-  text-transform: uppercase;
-  letter-spacing: 0.16em;
-  color: var(--ustock-text-muted);
-}
-
-.stat-value {
-  font-size: 24px;
-  font-weight: 700;
-  margin: 6px 0;
-  color: var(--ustock-text-strong);
-}
-
-.stat-caption {
-  font-size: 12px;
-  color: var(--ustock-text-muted);
-}
-
-.stat-card--clickable {
-  cursor: pointer;
-  text-decoration: none;
-  color: inherit;
-  transition: transform 0.15s ease, box-shadow 0.15s ease;
-}
-
-.stat-card--clickable:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 10px 25px rgba(15, 23, 42, 0.12);
-}
-
-/* Dark-mode slightly stronger shadow so it still “pops” */
-body.ustock-dark .stat-card--clickable:hover {
-  box-shadow: 0 12px 32px rgba(0, 0, 0, 0.7);
-}
-
-/* Mobile: stack the 3 stat cards */
-@media (max-width: 768px) {
-  .stat-row {
-    grid-template-columns: 1fr;
-  }
-}
