--- src/components/layout/AppShell.css
@@ -1,223 +1,0 @@
-:root {
-  /* THEME TOKENS */
-  --ustock-green-main: #0f766e;
-  --ustock-green-soft: #bbf7d0;
-
-  /* light theme defaults */
-  --ustock-bg-page: #f5f7ff;
-  --ustock-bg-card: #ffffff;
-  --ustock-bg-card-soft: #f9fafb;
-  --ustock-border-card: rgba(148, 163, 184, 0.25);
-
-  --ustock-text-main: #0f172a;
-  --ustock-text-strong: #0b1120;
-  --ustock-text-muted: #6b7280;
-
-  --ustock-card-chip-bg: #ecfdf5;
-}
-
-/* Base document background */
-html,
-body {
-  margin: 0;
-  padding: 0;
-  font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI",
-    sans-serif;
-  background: var(--ustock-bg-page);
-  color: var(--ustock-text-main);
-}
-
-/* ========= APP SHELL LAYOUT ======================================== */
-
-.app-shell {
-  min-height: 100vh;
-  background: var(--ustock-bg-page);
-  display: flex;
-  position: relative;
-  overflow-x: hidden;
-}
-
-/* when navOpen = true, margin-left is added to content (desktop only) */
-.app-shell--nav-open .app-main {
-  margin-left: 260px;
-}
-
-.app-main {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  transition: margin-left 0.23s ease-out;
-
-  /* 👇 make room for fixed topbar */
-  padding-top: 64px;
-}
-
-/* Wraps each page's content area */
-.app-content {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-}
-
-/* Shared page wrapper: controls max width + padding for ALL pages */
-.app-page {
-  width: 100%;
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 24px 48px 48px; /* top | sides | bottom */
-  box-sizing: border-box;
-}
-
-/* Tablets / small laptops */
-@media (max-width: 1024px) {
-  .app-page {
-    padding: 20px 24px 40px;
-  }
-}
-
-/* Mobile */
-@media (max-width: 768px) {
-  .app-page {
-    padding: 16px 24px 32px;
-  }
-
-  .app-shell--nav-open .app-main {
-    margin-left: 0;
-  }
-
-  /* slightly smaller gap on small screens if you want */
-  .app-main {
-    padding-top: 60px;
-  }
-}
-
-/* Extra-small mobile: make the main column fill the viewport exactly */
-@media (max-width: 480px) {
-  .app-page {
-    max-width: none;          /* stop trying to center a 1280px column */
-    width: 100vw;             /* match the viewport width */
-    margin: 0;                /* no auto-centering margins */
-    padding: 16px 16px 32px;  /* symmetric side padding */
-    box-sizing: border-box;
-  }
-}
-
-/* -------- FOOTER --------------------------------------------------- */
-
-.site-footer {
-  margin-top: auto;
-  text-align: center;
-  font-size: 12px;
-  color: var(--ustock-text-muted);
-  padding: 14px 0 0;
-  opacity: 0.9;
-}
-
-/* -------- GLOBAL UTILITIES ---------------------------------------- */
-
-.page-title {
-  color: var(--ustock-green-main);
-  font-weight: 800;
-  margin: 0;
-}
-
-.muted {
-  color: var(--ustock-text-muted);
-}
-
-/* Base “panel/card” style used across the app */
-.panel {
-  background: var(--ustock-bg-card);
-  border-radius: 18px;
-  padding: 18px 20px;
-  box-shadow: 0 10px 30px rgba(15, 23, 42, 0.06);
-  border: 1px solid var(--ustock-border-card);
-  color: var(--ustock-text-main);
-}
-
-/* Shared title + text helpers other pages can reuse */
-.panel-title {
-  font-size: 15px;
-  font-weight: 600;
-  color: var(--ustock-text-strong);
-  margin: 0;
-}
-
-.panel-subtitle {
-  font-size: 12px;
-  color: var(--ustock-text-muted);
-  margin: 0;
-}
-
-/* Simple muted mini list used across dashboards */
-.mini-list {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  font-size: 12px;
-}
-
-.mini-list li + li {
-  margin-top: 4px;
-}
-
-/* ========= LIGHT / DARK THEME TOKENS =============================== */
-
-/* light theme is already encoded in :root; just restate for clarity */
-body {
-  background: var(--ustock-bg-page);
-  color: var(--ustock-text-main);
-}
-
-/* Override tokens when the dark class is present */
-body.ustock-dark {
-  background: #020617;
-  color: #e5e7eb;
-
-  --ustock-bg-page: #020617;
-  --ustock-bg-card: #020617;
-  --ustock-bg-card-soft: #020617;
-  --ustock-border-card: rgba(148, 163, 184, 0.4);
-
-  --ustock-text-main: #e5e7eb;
-  --ustock-text-strong: #f9fafb;
-  --ustock-text-muted: #9ca3af;
-
-  --ustock-card-chip-bg: rgba(15, 118, 110, 0.25);
-}
-
-/* main layout background tweak for dark */
-body.ustock-dark .app-shell {
-  background: #020617;
-}
-
-body.ustock-dark .app-main {
-  background: radial-gradient(
-    circle at top left,
-    #022c22 0%,
-    #020617 40%,
-    #020617 100%
-  );
-}
-
-/* cards / panels / settings form – extra depth in dark */
-body.ustock-dark .panel,
-body.ustock-dark .about-card,
-body.ustock-dark .about-hero-card,
-body.ustock-dark .settings-card {
-  box-shadow: 0 18px 45px rgba(0, 0, 0, 0.6);
-}
-
-body.ustock-dark .settings-input,
-body.ustock-dark .settings-select,
-body.ustock-dark .settings-textarea {
-  background: var(--ustock-bg-card);
-  color: var(--ustock-text-main);
-  border-color: rgba(148, 163, 184, 0.7);
-}
-
-/* footer in dark */
-body.ustock-dark .site-footer {
-  color: #9ca3af;
-}
