+++ src/css/common/SearchableTickerDropdown.css
