+++ src/css/dashboard/cards/StatSummary.css
